+++ NEWS.md
@@ -2,6 +2,7 @@ Version 2.0.6
 =============
 
 * Added cloglog link for binomial, ZIB and ZNIB.
+* Added negative binomial (1) (VA via PIG augmentation)
 
 ## Bugfixes